tests: Avoid build step by using karma ts loader
