+++ tests/observing-keys.spec.ts
@@ -1,22 +1,19 @@
 import * as m from 'mochainon';
 
 import * as localforage from 'localforage';
-import '../../';
+import '../';
 
 import { formatChangeArg } from './utils/formatChangeArg';
 const { expect } = m.chai;
 
-describe('Observing methods', function() {
-    let setItemSubscription: Subscription;
-    let clearSubscription: Subscription;
-    let setItemObservableLogs: string[];
-    let clearObservableLogs: string[];
+describe('Observing keys', function() {
+    let subscription: Subscription;
+    let observableLogs: string[];
     let errorCallCount: number;
     let completeCallCount: number;
 
     beforeEach(function() {
-        setItemObservableLogs = [];
-        clearObservableLogs = [];
+        observableLogs = [];
         errorCallCount = 0;
         completeCallCount = 0;
     });
@@ -52,35 +49,18 @@ describe('Observing methods', function() {
             // this should not notify subscribers w/ changeDetection
             .then(() => localforage.clear());
 
-    describe('Given a setItem & a clear observable & subscription w/ changeDetection', function() {
+    describe('Given a simple observable & subscription w/ changeDetection', function() {
         beforeEach(function() {
             return localforage
                 .ready()
                 .then(() => localforage.clear())
                 .then(() => {
-                    const setItemCallObservable = localforage.newObservable({
-                        setItem: true,
+                    const observable = localforage.newObservable({
+                        key: 'UserProfile',
                     });
-
-                    setItemSubscription = setItemCallObservable.subscribe({
+                    subscription = observable.subscribe({
                         next: (change: LocalForageObservableChange) =>
-                            setItemObservableLogs.push(formatChangeArg(change)),
-                        error: err => {
-                            errorCallCount++;
-                            console.error('Found an error!', err);
-                        },
-                        complete: () => {
-                            completeCallCount++;
-                        },
-                    });
-
-                    const clearCallObservable = localforage.newObservable({
-                        clear: true,
-                    });
-
-                    clearSubscription = clearCallObservable.subscribe({
-                        next: (change: LocalForageObservableChange) =>
-                            clearObservableLogs.push(formatChangeArg(change)),
+                            observableLogs.push(formatChangeArg(change)),
                         error: err => {
                             errorCallCount++;
                             console.error('Found an error!', err);
@@ -95,60 +75,41 @@ describe('Observing methods', function() {
         it('should observe properly', function() {
             return runTestScenario()
                 .then(function() {
-                    setItemSubscription.unsubscribe();
-                    clearSubscription.unsubscribe();
+                    subscription.unsubscribe();
+                    return localforage.setItem(
+                        'notObservedKey',
+                        'notObservedValue',
+                    );
+                })
+                .then(function() {
                     return localforage.clear();
                 })
                 .then(function() {
-                    expect(setItemObservableLogs).to.deep.equal([
+                    expect(observableLogs).to.deep.equal([
                         `setItem('UserProfile', '{"UserName":"user1","Password":"12345"}')`,
                         `setItem('UserProfile', '{"UserName":"user1","Password":"67890"}')`,
-                        `setItem('test1', 'value1')`,
-                        `setItem('test2', 'value2')`,
-                        `setItem('test2', 'value2b')`,
-                        `setItem('test3', 'value3')`,
+                        // TODO: fix me
+                        // 'clear()',
                     ]);
-                    expect(clearObservableLogs)
-                        // TODO: Fix this
-                        // .to.deep.equal(['clear()']);
-                        .to.deep.equal([]);
                     expect(errorCallCount).to.equal(0);
                     expect(completeCallCount).to.equal(0);
                 });
         });
     });
 
-    describe('Given a setItem & a clear observable & subscription w/o changeDetection', function() {
+    describe('Given a simple observable & subscription w/o changeDetection', function() {
         beforeEach(function() {
             return localforage
                 .ready()
                 .then(() => localforage.clear())
                 .then(() => {
-                    const setItemCallObservable = localforage.newObservable({
-                        setItem: true,
+                    const observable = localforage.newObservable({
+                        key: 'UserProfile',
                         changeDetection: false,
                     });
-
-                    setItemSubscription = setItemCallObservable.subscribe({
+                    subscription = observable.subscribe({
                         next: (change: LocalForageObservableChange) =>
-                            setItemObservableLogs.push(formatChangeArg(change)),
-                        error: err => {
-                            errorCallCount++;
-                            console.error('Found an error!', err);
-                        },
-                        complete: () => {
-                            completeCallCount++;
-                        },
-                    });
-
-                    const clearCallObservable = localforage.newObservable({
-                        clear: true,
-                        changeDetection: false,
-                    });
-
-                    clearSubscription = clearCallObservable.subscribe({
-                        next: (change: LocalForageObservableChange) =>
-                            clearObservableLogs.push(formatChangeArg(change)),
+                            observableLogs.push(formatChangeArg(change)),
                         error: err => {
                             errorCallCount++;
                             console.error('Found an error!', err);
@@ -163,24 +124,22 @@ describe('Observing methods', function() {
         it('should observe properly', function() {
             return runTestScenario()
                 .then(function() {
-                    setItemSubscription.unsubscribe();
-                    clearSubscription.unsubscribe();
+                    subscription.unsubscribe();
+                    return localforage.setItem(
+                        'notObservedKey',
+                        'notObservedValue',
+                    );
+                })
+                .then(function() {
                     return localforage.clear();
                 })
                 .then(function() {
-                    expect(setItemObservableLogs).to.deep.equal([
+                    expect(observableLogs).to.deep.equal([
                         `setItem('UserProfile', '{"UserName":"user1","Password":"12345"}')`,
                         `setItem('UserProfile', '{"UserName":"user1","Password":"67890"}')`,
                         `setItem('UserProfile', '{"UserName":"user1","Password":"67890"}')`,
-                        `setItem('test1', 'value1')`,
-                        `setItem('test2', 'value2')`,
-                        `setItem('test2', 'value2b')`,
-                        `setItem('test2', 'value2b')`,
-                        `setItem('test3', 'value3')`,
-                    ]);
-                    expect(clearObservableLogs).to.deep.equal([
-                        'clear()',
-                        'clear()',
+                        // TODO: fix me
+                        // 'clear()',
                     ]);
                     expect(errorCallCount).to.equal(0);
                     expect(completeCallCount).to.equal(0);
